ril: Only include libsecril-client if TARGET_AUDIOHAL_VARIANT is set
After moving the OSS libsecril-client to vendor, there's clashing with
the proprietary libsecril-client blobs that some devices have.

The OSS libsecril-client is only usable when using the OSS audio HAL
here, which requires setting TARGET_AUDIOHAL_VARIANT. So, include it
only if TARGET_AUDIOHAL_VARIANT is set at all (to 'samsung' or any other
value for a custom copy of the HAL).

Change-Id: I320636ed0b2810c727d68fed493caae1c4893b5b